Include thrown exception in onError() callback
Until now, it was not possible to know the reason of an error during RSS parsing because exceptions where ignored. Now, the onError() method includes the exception that was thrown so that the developer can investigate further.
Note that this is a breaking change, althought it shouldn't be difficult to adapt.